+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Added
 - Parsing of `variantType`, `referenceCopyNumber`, `copyNumber` and `chromosomeCoordinates` fields from CSV/TSV files
 - Genome assembly can be passed as an optional param to tsv_2_json and csv_to_json to create submission containing chromosome coordinates
+### Fixed
+- Parsing of multiple condition IDs from csv/tsv files
 
 ## [2.0.1]
 ### Fixed